feat: Day 3 - Parsing the new endpoint:
{item_id}: Path parameter.
item_id: int: Here comes the magic. We specify that item_id should be an integer (int). FastAPI will:
Validate the incoming data. Try passing abc in the documentation - you will get a nice 422 Unprocessable Entity error.
Convert the type. The string 123 from the URL will be automatically converted to the number 123 in your function.
Document this requirement in /docs.